ispa/requirements/base.txt
psycopg2==2.7.1
Werkzeug
wheel==0.29.0
django==1.10.7
django-environ==0.4.3
django-model-utils==3.0.0
Pillow==4.2.1
awesome-slugify==1.6.5
djangorestframework==3.6.3
django-rest-auth[with_social]
markdown==2.6.8
wagtail==1.11.1
argon2-cffi==16.3.0
graphene-django>=1.0
coreapi==2.3.1
